Handle whitespace inside closing tags
Fixes #65 by adding a skip whitespace before and after the name of a
closing tag. Also adds a test to verify that this functions properly,
based on the example in #65.

Note that the [HTML spec] does not allow for whitespace between the `</`
and the tag name, although this is accepted by these changes to maintain
symmetry with the parsing of the start of a tag which started allowing
whitespace between `<` and the tag name in c532c72.

[HTML spec]: https://html.spec.whatwg.org/multipage/syntax.html#end-tags